Refactor, Cleanup and Consolidate
- Put OpModes into packages
- disable unused OpModes
- Create new "CompetionBot" with OpModes - combined Jay and Kerric code into these
