+++ ftc_app-master_tb/TeamCode/src/main/java/org/firstinspires/ftc/teamcode/testbot/TestBot_Drive.java
@@ -1,27 +1,21 @@
 
 
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.testbot;
 
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
 
-@TeleOp(name="TestBot: DriveAndLiftWithClaw", group="test")
+@TeleOp(name="TestBot: Drive", group="test")
 
-public class TestBot_DriveAndLiftWithClaw extends OpMode{
+public class TestBot_Drive extends OpMode{
 
     /* Declare OpMode members. */
     TestBot robot       = new TestBot(telemetry);
 
-
-    double          clawOffset      = 0;                       // Servo mid position
-    final double    CLAW_SPEED      = 0.02 ;                   // sets rate to move servo
-
     /*
      * Code to run ONCE when the driver hits INIT
      */
-
-
     @Override
     public void init() {
         /* Initialize the hardware variables.
@@ -69,26 +63,6 @@ public class TestBot_DriveAndLiftWithClaw extends OpMode{
         telemetry.addData("rightX", "%.2f", rightX);
         telemetry.addData("rightY", "%.2f", rightY);
 
-        // move arm up if a button pushed
-        if(gamepad1.right_trigger > 0){
-            robot.leftArm.setPower(-0.5);
-
-        // move arm down if b button pushed
-        } else if(gamepad1.left_trigger > 0){
-            robot.leftArm.setPower(.5);
-
-        // if neither button pushed, stop arm
-        } else {
-            robot.leftArm.setPower(0);
-        }
-
-        // Use dpad to open and close the claw
-        if (gamepad1.dpad_up)
-            clawOffset += CLAW_SPEED;
-        else if (gamepad1.dpad_down)
-            clawOffset -= CLAW_SPEED;
-
-        robot.moveClaws(clawOffset);
     }
 
     /*
