+++ ftc_app-master_tb/TeamCode/src/main/java/sample_camera_opmodes/DetectColor.java
@@ -2,6 +2,7 @@ package sample_camera_opmodes;
 
 import android.graphics.Bitmap;
 
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
 import for_camera_opmodes.OpModeCamera;
@@ -13,7 +14,8 @@ import for_camera_opmodes.OpModeCamera;
  */
 
 @TeleOp(name = "DetectColor", group = "ZZOpModeCameraPackage")
-//@Disabled
+@Disabled
+
 public class DetectColor extends OpModeCamera {
 
   int ds2 = 8;  // additional downsampling of the image
